+++ hyperdbg/hyperhv/header/assembly/InlineAsm.h
@@ -97,6 +97,12 @@ AsmVmexitHandler();
  */
 extern void inline AsmSaveVmxOffState();
 
+/**
+ * @brief Restore XMM registers
+ *
+ */
+extern void inline AsmVmxoffRestoreXmmRegs(unsigned long long XmmRegs);
+
 //
 // ====================  Extended Page Tables ====================
 // File : AsmEpt.asm
@@ -143,7 +149,7 @@ AsmGetCs();
 extern unsigned short
 AsmGetDs();
 
-extern void 
+extern void
 AsmSetDs(unsigned short DsSelector);
 
 /**
@@ -281,6 +287,13 @@ AsmReloadGdtr(void * GdtBase, unsigned long GdtLimit);
  */
 extern void
 AsmReloadIdtr(void * GdtBase, unsigned long GdtLimit);
+
+/**
+ * @brief Read SSP
+ *
+ */
+extern unsigned long long
+AsmReadSsp();
 
 //
 // ====================  Hook Functions ====================