+++ hyperdbg/include/SDK/headers/BasicTypes.h
@@ -66,6 +66,7 @@ typedef unsigned __int64 UINT64, *PUINT64;
 //
 // DO NOT FUCKING TOUCH THIS STRUCTURE WITHOUT COORDINATION WITH SINA
 //
+#pragma pack(push, 1) // Ensure no padding
 typedef struct GUEST_REGS
 {
     //
@@ -94,6 +95,38 @@ typedef struct GUEST_REGS
     //
 
 } GUEST_REGS, *PGUEST_REGS;
+#pragma pack(pop)
+
+typedef struct XMM_REG
+{
+    UINT64 XmmLow;  // Low 64 bits
+    UINT64 XmmHigh; // High 64 bits
+
+} XMM_REG;
+
+#pragma pack(push, 1) // Ensure no padding
+typedef struct GUEST_XMM_REGS
+{
+    XMM_REG xmm0;  // 0x00
+    XMM_REG xmm1;  // 0x10
+    XMM_REG xmm2;  // 0x20
+    XMM_REG xmm3;  // 0x30
+    XMM_REG xmm4;  // 0x40
+    XMM_REG xmm5;  // 0x50
+    XMM_REG xmm6;  // 0x60
+    XMM_REG xmm7;  // 0x70
+    XMM_REG xmm8;  // 0x80
+    XMM_REG xmm9;  // 0x90
+    XMM_REG xmm10; // 0xA0
+    XMM_REG xmm11; // 0xB0
+    XMM_REG xmm12; // 0xC0
+    XMM_REG xmm13; // 0xD0
+    XMM_REG xmm14; // 0xE0
+    XMM_REG xmm15; // 0xF0
+    UINT32  mxcsr; // 0x100
+
+} GUEST_XMM_REGS, *PGUEST_XMM_REGS;
+#pragma pack(pop)
 
 /**
  * @brief struct for extra registers