+++ hyperdbg/libhyperdbg/header/commands.h
@@ -472,6 +472,9 @@ typedef std::map<std::string, COMMAND_DETAIL> CommandType;
 #define DEBUGGER_COMMAND_IDT_ATTRIBUTES \
     DEBUGGER_COMMAND_ATTRIBUTE_LOCAL_COMMAND_IN_DEBUGGER_MODE
 
+#define DEBUGGER_COMMAND_SMI_ATTRIBUTES \
+    DEBUGGER_COMMAND_ATTRIBUTE_LOCAL_COMMAND_IN_DEBUGGER_MODE
+
 //////////////////////////////////////////////////
 //             Command Functions                //
 //////////////////////////////////////////////////
@@ -754,6 +757,9 @@ CommandPcicam(vector<CommandToken> CommandTokens, string Command);
 
 VOID
 CommandIdt(vector<CommandToken> CommandTokens, string Command);
+
+VOID
+CommandSmi(vector<CommandToken> CommandTokens, string Command);
 
 //
 // hwdbg commands